Move serde codegen / derive stuff to its own file
This matches the Serde docs from https://serde.rs/codegen-stable.html
and should be more discoverable for new contributors.

We should also get slightly better error messages now that the generated
code is more isolated.
