+++ build.rs
@@ -1,5 +1,7 @@
-// Stable Rust 1.9 doesn't support the custom_derive feature, so we generate
+// Stable Rust doesn't yet support the custom_derive feature, so we generate
 // code for structures which #[derive] Serde's Serialize/Deserialize traits.
+//
+// See https://serde.rs/codegen-stable.html for more information.
 
 extern crate serde_codegen;
 
@@ -9,8 +11,8 @@ use std::path::Path;
 pub fn main() {
     let out_dir = env::var_os("OUT_DIR").unwrap();
 
-    let src = Path::new("src/lib.rs.in");
-    let dst = Path::new(&out_dir).join("lib.rs");
+    let src = Path::new("src/serde_types.in.rs");
+    let dst = Path::new(&out_dir).join("serde_types.rs");
 
     serde_codegen::expand(&src, &dst).unwrap();
 }
