+++ src/lib.rs
@@ -4,6 +4,7 @@ extern crate log;
 extern crate hyper;
 extern crate iron;
 extern crate redis;
+extern crate serde;
 extern crate serde_json;
 extern crate time;
 extern crate url;
